Remove FormValidationDirective, introduce FormUtilsService
The FormValidationDirective only had one task: marking all fields as dirty on submit.
This was done in an implicit way and hidden in the directive.
I went for a rather explicit way now. This commit introduces the FormUtilsService that, in the moment, only has one method `markAsDirtyRecursive()`.
This method is called in the submit handler for each form involved. Thus, you can easily see that the "dirty marking" is done on submit (or not).
The FormValidationDirective is being removed.