Complete rework of registration page
This is a very huge commit that will not be easy to merge. 💥 Sorry...
However, it introduces some structural reconsideration that hopefully makes the application more future-proof. 🚀

TL;DR
- rework of registration page
- address forms work by new architecture
- strict split by presentational/container components
- everything is top-down now
- validation is in there
- minor stuff

## Introduce AddressFormModule
- is a refactored version of the old address forms from the ExtraFormsModule
- all AddressXX forms and their FormGroup factories
- AddressFormService with utils

## Remove ExtraFormsModule
All controls temporarily went to the SharedModule – temporarily!
We need to refactor this later.

## Revise Registration Page
This is a complete refactor of the registration page.
It is built in a new folder "registration-page-revised" in order to not completely break things.

- RegistrationPage is the container component (the only component that communicates with services and the outside world)
- all other components are strict presentational components (if there is service communication, then only with local services like TranslateService, FormUtilsService and AdressFormService)
- this architecture is more clean and even necessary when we move to ngrx

- All data walks top-down!
- Service communication is made on top and then the data is being passed downwards to the children (see regions, languages and countries for that).
- The complete FormGroup lives in the top component "RegistrationForm". That group is being passed down and all child components only access the controls from there.

- Form Validation with the newly created components and directive is built in everywhere now (hopefully everywhere).

## CountryService / RegionService
The Observables now return one single item with an array of countries/regions.
This is a more real-world behavior since the REST service also returns an array in a single response.

## Specific Select components
There are some specific select form components (select-language, select-region, select-country).
They were all inherited from SelectComponent while the template also came from there. Only the logic was being extended.

This architecture changed now!
The specific select components now only prepare data and set defaults – and pass them down to a real SelectComponent which resides in their view. This is a more decoupled way of specializing this.

## Misc
- no tests touched...
- some minor changes wherever I ran into them
- merging will be fun... 😇