feat: add configurable product attributes
An attempt to make product defaut easily overridable was mande on ticket ##1098
However, this change
only affects the product listing.
It is not possible to easily add an attribute globally on all the
products of the site (and the api rest on the backoffice side is quite rigid).
This modification
introduces a service which allows to solve this problem more easily and also to make this
configuration parameterizable at the level of an extension by using the \"FeatureToggleService\".

BREAKING CHANGES:
products-list-attribute.ts no longer provide default export