fix: give the session a chance to be refreshed
After the 3.2.0 refactoring, the session token is refreshed via an SSO mechanism. However, the
expiration time of the OAuth token is the same as that of the session cookie. When the refresh time
arrives, the cookie expires before the refresh token request can be made and the session is
destroyed.
An extra minute of session time is given to the cookie to avoid this bug