fix: force a name attribute on radio inputs
Radio inputs now have a correctly filled name attribute.
When this attribute is not set and we bind
a [formGroup] on the input, angular considers that the radio button is global.
It is therefore
impossible to create several different groups of radio buttons on the page (changing one of the
buttons will erase the status of the others).
This fix prevents this unexpected behavior