fix: parameters ;loc and ;cur may overlook server configuration
When the first page of the site is displayed, requests can be sent to the icm before the redux
ServerConfiguration store is initialized. This can be problematic in the case of localized URLs such
as content retrieval, as the icm channel configuration can influence the value of the ;loc and ;cur
parameters.