fix: fix: handle token errors for local development
In the case of local development, you can change the target back-end by modifying the value of the
variable icmBaseURL' variable in environment.developement.ts. By doing so, you will need a new api
token. However, as your local DNS is still \"localhost\", there is no token renewal and the old token
is kept. This small modification corrects this problem.